SyncServerWalDump/requirements.txt
pkg-resources==0.0.0
psycopg2-binary==2.8.4
